Refactor: libcrmcommon: separate schema-related code into its own source file
At nearly 6,000 lines, xml.c was difficult to browse. Schema-aware functions
are higher-level than simple XML functions, so it makes sense to isolate them.
